requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-filter==23.3
cx_Oracle==8.3.0
reportlab==4.0.7
matplotlib==3.8.2
Pillow==10.1.0
Faker==20.1.0
django-cors-headers==4.3.1

